% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power}
\alias{power}
\title{power calculation in one-sample log-rank test}
\usage{
power(alpha, n, ta, tf, m0, delta, k)
}
\arguments{
\item{alpha}{type I error rate}

\item{n}{sample size}

\item{ta}{length of accrual period, during which patients are recruited}

\item{tf}{length of follow-up time, during which patients are monitored}

\item{m0}{median survival time of the historical control sample, which shall be obtained from previous research.}

\item{delta}{expected effect size.}

\item{k}{shape parameter of survival functions (assumed to be Weibull).}
}
\value{
power in one-sample log-rank test
}
\description{
calculate power in one-sample log-rank test
}
\examples{
power(alpha = 0.05, n = 88,  ta = 5, tf = 3, m0 = 9, delta = 4/7, k = 1.22)
}
