% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSize.R
\name{SampleSize}
\alias{SampleSize}
\title{sample size calculation in one-sample log-rank test}
\usage{
SampleSize(alpha, power, ta, tf, m0, delta, k)
}
\arguments{
\item{alpha}{type I error rate}

\item{power}{desired power for your study}

\item{ta}{length of accrual period, during which patients are recruited}

\item{tf}{length of follow-up time, during which patients are monitored}

\item{m0}{median survival time of the historical control sample, which shall be obtained from previous research.}

\item{delta}{expected effect size.}

\item{k}{shape parameter of survival functions (assumed to be Weibull).}
}
\value{
sample size in one-sample log-rank test
}
\description{
calculate sample size in one-sample log-rank test
}
\examples{
SampleSize(alpha = 0.05, power = 0.8, ta = 5, tf = 3, m0 = 9, delta = 0.57, k = 1.22)
}
