% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{Simulation}
\alias{Simulation}
\title{empirical type I error and power}
\usage{
Simulation(n, parameter, B = 10000, k, delta, m0, ta, tf)
}
\arguments{
\item{n}{sample size}

\item{parameter}{if true, calculate alpha, otherwise calculate power}

\item{B}{number of iterations used in simulation}

\item{k}{shape parameter of survival functions (assumed to be Weibull).}

\item{delta}{expected effect size.}

\item{m0}{median survival time of the historical control sample, which shall be obtained from previous research.}

\item{ta}{length of accrual period, during which patients are recruited}

\item{tf}{length of follow-up time, during which patients are monitored}
}
\value{
type I error or power in one-sample log-rank test
}
\description{
calculate empirical type I error and power in one-sample log-rank test
}
\examples{
Simulation(n = 534, parameter = TRUE, B = 10000, ta = 3, tf = 1, m0 = 1, delta = 1/1.2, k = 0.1)
Simulation(n = 534, parameter = FALSE, B = 10000, ta = 3, tf = 1, m0 = 1, delta = 1/1.2, k = 0.1)
}
